+++ LifeLineDemo.Infrastructure/Queries/GetHospital.cs
@@ -2,11 +2,6 @@
 using LifeLineDemo.Domain.DTO;
 using LifeLineDemo.Domain.Interface.Queries;
 using LifeLineDemo.Infrastructure.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace LifeLineDemo.Infrastructure.Queries
 {
@@ -15,12 +10,12 @@ namespace LifeLineDemo.Infrastructure.Queries
         private readonly LifeLineDbContext lifeLineDb;
         private readonly IMapper mapper;
 
-        public GetHospital(LifeLineDbContext lifeLineDb,IMapper mapper)
+        public GetHospital(LifeLineDbContext lifeLineDb, IMapper mapper)
         {
             this.lifeLineDb = lifeLineDb;
             this.mapper = mapper;
         }
-        public HospitalDto GetHospitalById(int id)
+        public HospitalDto GetHospitalById(long id)
         {
             var hospital = lifeLineDb.Hospitals.FirstOrDefault(x => x.Id == id);
             return mapper.Map<HospitalDto>(hospital);
@@ -28,7 +23,7 @@ namespace LifeLineDemo.Infrastructure.Queries
 
         public IList<HospitalDto> GetHospitalList()
         {
-            var hospitallist=lifeLineDb.Hospitals.ToList();
+            var hospitallist = lifeLineDb.Hospitals.ToList();
             return mapper.Map<IList<HospitalDto>>(hospitallist);
         }
     }
